Create AvatarStateExtensions
